Move EscapeHtml helper to escape HTML reserved characters to a separate namespace
